3. Neither irony nor sarcasm is argument

+++ 0x0A-argc_argv/3-mul.c
@@ -14,7 +14,7 @@
 int main(int argc, char *argv[])
 {
 	int i;
-	float mul = 1;
+	double mul = 1;
 
 	if (argc == 1)
 	{
